# mobile/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "每日減脂計畫 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/foods": {
      "get": {
        "tags": [
          "foods"
        ],
        "summary": "Search Foods",
        "operationId": "search_foods_foods_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FoodOut"
                  },
                  "title": "Response Search Foods Foods Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "foods"
        ],
        "summary": "Add Food",
        "operationId": "add_food_foods_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/foods/{food_id}": {
      "get": {
        "tags": [
          "foods"
        ],
        "summary": "Read Food",
        "operationId": "read_food_foods__food_id__get",
        "parameters": [
          {
            "name": "food_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Food Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "foods"
        ],
        "summary": "Edit Food",
        "operationId": "edit_food_foods__food_id__patch",
        "parameters": [
          {
            "name": "food_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Food Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "foods"
        ],
        "summary": "Remove Food",
        "operationId": "remove_food_foods__food_id__delete",
        "parameters": [
          {
            "name": "food_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Food Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/foods/{food_id}/exchanges": {
      "get": {
        "tags": [
          "foods"
        ],
        "summary": "List Exchanges",
        "description": "Same-category swaps. `grams` defaults to the food's usual portion, `match` to\nthe category's own rule.",
        "operationId": "list_exchanges_foods__food_id__exchanges_get",
        "parameters": [
          {
            "name": "food_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Food Id"
            }
          },
          {
            "name": "grams",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grams"
            }
          },
          {
            "name": "match",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Match"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExchangeOut"
                  },
                  "title": "Response List Exchanges Foods  Food Id  Exchanges Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExchangeOut": {
        "properties": {
          "food": {
            "$ref": "#/components/schemas/FoodOut"
          },
          "grams": {
            "type": "number",
            "title": "Grams"
          },
          "nutrients": {
            "$ref": "#/components/schemas/NutrientsOut"
          },
          "delta": {
            "$ref": "#/components/schemas/NutrientsOut"
          },
          "capped": {
            "type": "boolean",
            "title": "Capped"
          }
        },
        "type": "object",
        "required": [
          "food",
          "grams",
          "nutrients",
          "delta",
          "capped"
        ],
        "title": "ExchangeOut"
      },
...
      "FoodIn": {
        "properties": {
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "type": "string",
            "enum": [
              "raw",
              "cooked",
              "na"
            ],
            "title": "State",
            "default": "na"
          },
          "kcal_per_100g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Kcal Per 100G"
          },
          "protein_per_100g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Protein Per 100G"
          },
          "fat_per_100g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Fat Per 100G"
          },
          "carb_per_100g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Carb Per 100G"
          },
          "fiber_per_100g": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiber Per 100G"
          },
          "unit": {
            "type": "string",
            "enum": [
              "g",
              "ml",
              "piece",
              "scoop",
              "bowl"
            ],
            "title": "Unit",
            "default": "g"
          },
          "grams_per_unit": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Grams Per Unit"
          },
          "usual_grams": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Usual Grams"
          },
          "max_grams": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Grams"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "name",
          "kcal_per_100g",
          "protein_per_100g",
          "fat_per_100g",
          "carb_per_100g",
          "usual_grams",
          "max_grams"
        ],
        "title": "FoodIn"
      },
      "FoodOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "per_100g": {
            "$ref": "#/components/schemas/NutrientsOut"
          },
          "fiber_per_100g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiber Per 100G"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "grams_per_unit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grams Per Unit"
          },
          "usual_grams": {
            "type": "number",
            "title": "Usual Grams"
          },
          "max_grams": {
            "type": "number",
            "title": "Max Grams"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "name",
          "state",
          "per_100g",
          "fiber_per_100g",
          "unit",
          "grams_per_unit",
          "usual_grams",
          "max_grams",
          "aliases",
          "template_id"
        ],
        "title": "FoodOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NutrientsOut": {
        "properties": {
          "kcal": {
            "type": "number",
            "title": "Kcal"
          },
          "protein_g": {
            "type": "number",
            "title": "Protein G"
          },
          "fat_g": {
            "type": "number",
            "title": "Fat G"
          },
          "carb_g": {
            "type": "number",
            "title": "Carb G"
          }
        },
        "type": "object",
        "required": [
          "kcal",
          "protein_g",
          "fat_g",
          "carb_g"
        ],
        "title": "NutrientsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}